きのこが設置中なら大きさが変化するように

+++ 3D_Base/Source/GameObject/MeshObject/StaticMesh/Item/Items/Bomb/Bomb.cpp
@@ -283,10 +283,10 @@ void Bomb::ChangeColor()
 	m_ColorTimer += CTimeManager::GetDeltaTime();
 
 	//点滅のスピードを経過時間/爆発するまでの時間をして割合で出す
-	float speed = 10.0f * (static_cast<float>((m_ColorTimer) / m_ExplosionTime));
+	float speed = 10.0f * (m_ColorTimer / m_ExplosionTime);
 
 	//+1.0fをすることで、sinの値が0~2の間の値になり、*0.5することで0~1の間の値が取れる
-	float blinkRate = (sinf(static_cast<float>(m_ColorTimer) * speed) + 1.0f) * 0.5f;
+	float blinkRate = (sinf(m_ColorTimer * speed) + 1.0f) * 0.5f;
 
 	//灰色のカラーコード
 	D3DXVECTOR4 gray = D3DXVECTOR4(0.2f, 0.2f, 0.2f, 1.0f);
